Inline modules in `polkadot-node-subsystem-types` crate
