--- polkadot/node/subsystem-types/src/errors.rs
@@ -1,42 +1,0 @@
-// Copyright 2020 Parity Technologies (UK) Ltd.
-// This file is part of Polkadot.
-
-// Polkadot is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Polkadot is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
-
-//! Error types for the subsystem requests.
-
-/// An error type that describes faults that may happen
-///
-/// These are:
-///   * Channels being closed
-///   * Subsystems dying when they are not expected to
-///   * Subsystems not dying when they are told to die
-///   * etc.
-#[derive(thiserror::Error, Debug)]
-#[allow(missing_docs)]
-pub enum SubsystemError {
-	#[error(transparent)]
-	QueueError(#[from] futures::channel::mpsc::SendError),
-
-	/// Generated by the `#[overseer(..)]` proc-macro
-	#[error(transparent)]
-	Generated(#[from] ::polkadot_overseer_gen::OverseerError),
-
-	/// Generated by the `#[overseer(..)]` proc-macro
-	#[error(transparent)]
-	RuntimeApi(#[from] sp_api::ApiError),
-}
-
-/// Ease the use of subsystem errors.
-pub type SubsystemResult<T> = Result<T, self::SubsystemError>;
