+++ polkadot/node/subsystem-types/src/lib.rs
@@ -29,8 +29,51 @@ use cirrus_node_primitives::ExecutorSlotInfo;
 use smallvec::SmallVec;
 pub use subspace_runtime_primitives::{BlockNumber, Hash};
 
-pub mod errors;
-pub mod messages;
+pub mod errors {
+	//! Error types for the subsystem requests.
+
+	/// An error type that describes faults that may happen
+	///
+	/// These are:
+	///   * Channels being closed
+	///   * Subsystems dying when they are not expected to
+	///   * Subsystems not dying when they are told to die
+	///   * etc.
+	#[derive(thiserror::Error, Debug)]
+	#[allow(missing_docs)]
+	pub enum SubsystemError {
+		#[error(transparent)]
+		QueueError(#[from] futures::channel::mpsc::SendError),
+
+		/// Generated by the `#[overseer(..)]` proc-macro
+		#[error(transparent)]
+		Generated(#[from] ::polkadot_overseer_gen::OverseerError),
+
+		/// Generated by the `#[overseer(..)]` proc-macro
+		#[error(transparent)]
+		RuntimeApi(#[from] sp_api::ApiError),
+	}
+
+	/// Ease the use of subsystem errors.
+	pub type SubsystemResult<T> = Result<T, self::SubsystemError>;
+}
+pub mod messages {
+	use cirrus_node_primitives::CollationGenerationConfig;
+	use sp_executor::{BundleEquivocationProof, FraudProof, InvalidTransactionProof};
+
+	/// Message to the Collation Generation subsystem.
+	#[derive(Debug)]
+	pub enum CollationGenerationMessage {
+		/// Initialize the collation generation subsystem
+		Initialize(CollationGenerationConfig),
+		/// Fraud proof needs to be submitted to primary chain.
+		FraudProof(FraudProof),
+		/// Bundle equivocation proof needs to be submitted to primary chain.
+		BundleEquivocationProof(BundleEquivocationProof),
+		/// Invalid transaction proof needs to be submitted to primary chain.
+		InvalidTransactionProof(InvalidTransactionProof),
+	}
+}
 
 /// How many slots are stack-reserved for active leaves updates
 ///
